imageDisk: split subclasses into their own source files and headers
Give every imageDisk subclass its own include/imagedisk_*.h (+ matching
src/ints/imagedisk_*.cpp where it has out-of-line methods), mirroring the
existing imagedisk_teledisk pattern. Split out imageDiskVFD, imageDiskD88,
imageDiskNFD, imageDiskEmptyDrive, imageDiskINT13Drive, imageDiskElToritoFloppy
(header-only), and imageDiskMSDOSBlockDevice, whose implementation also moves
out of src/misc/programs.cpp. The base imageDisk class and BIOS disk-list
plumbing stay in bios_disk.{h,cpp}.

Pure reorganization with no behavior change.
