+++ include/bios_disk.h
@@ -191,121 +191,6 @@ class imageDisk {
 		}
 };
 
-class imageDiskEmptyDrive : public imageDisk {
-public:
-	Int13Status Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size=0) override;
-	Int13Status Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size=0) override;
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	imageDiskEmptyDrive();
-	virtual ~imageDiskEmptyDrive();
-};
-
-class imageDiskD88 : public imageDisk {
-public:
-	Int13Status Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size=0) override;
-	Int13Status Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size=0) override;
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	imageDiskD88(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk);
-	virtual ~imageDiskD88();
-
-    unsigned char fd_type_major;
-    unsigned char fd_type_minor;
-
-    enum { /* major */
-        DISKTYPE_2D			= 0,
-        DISKTYPE_2DD,
-        DISKTYPE_2HD
-    };
-
-    struct vfdentry {
-        uint8_t         track = 0,head = 0,sector = 0;
-        uint16_t        sector_size = 0;
-
-        uint32_t        data_offset = 0;
-        uint32_t        entry_offset = 0; // offset of the 12-byte entry this came from (if nonzero)
-
-        vfdentry() {}
-
-        uint16_t getSectorSize(void) const {
-            return sector_size;
-        }
-    };
-
-    vfdentry *findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size=0);
-
-    std::vector<vfdentry> dents;
-};
-
-class imageDiskNFD : public imageDisk {
-public:
-	Int13Status Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size=0) override;
-	Int13Status Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size=0) override;
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	imageDiskNFD(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk, unsigned int revision);
-	virtual ~imageDiskNFD();
-
-    struct vfdentry {
-        uint8_t         track = 0,head = 0,sector = 0;
-        uint16_t        sector_size = 0;
-
-        uint32_t        data_offset = 0;
-        uint32_t        entry_offset = 0; // offset of the 12-byte entry this came from (if nonzero)
-
-        vfdentry() {}
-
-        uint16_t getSectorSize(void) const {
-            return sector_size;
-        }
-    };
-
-    vfdentry *findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size=0);
-
-    std::vector<vfdentry> dents;
-};
-
-class imageDiskVFD : public imageDisk {
-public:
-	Int13Status Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size=0) override;
-	Int13Status Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size=0) override;
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	imageDiskVFD(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk);
-	virtual ~imageDiskVFD();
-
-    struct vfdentry {
-        uint8_t         track = 0,head = 0,sector = 0,sizebyte = 0;
-        uint8_t         fillbyte = 0xFF;
-
-        uint32_t        data_offset = 0;
-        uint32_t        entry_offset = 0; // offset of the 12-byte entry this came from (if nonzero)
-
-        vfdentry() {}
-
-        bool hasSectorData(void) const {
-            return fillbyte == 0xFF && data_offset != 0xFFFFFFFFUL;
-        }
-
-        bool hasFill(void) const {
-            return fillbyte != 0xFF || data_offset == 0xFFFFFFFFUL;
-        }
-
-        uint16_t getSectorSize(void) const {
-            return 128 << sizebyte;
-        }
-    };
-
-    vfdentry *findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size=0);
-
-    std::vector<vfdentry> dents;
-};
-
 class imageDiskMemory : public imageDisk {
 public:
 	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
@@ -480,87 +365,6 @@ private:
     //uint64_t image_length = 0;
 };
 
-/* C++ class implementing El Torito floppy emulation */
-class imageDiskElToritoFloppy : public imageDisk {
-public:
-    /* Read_Sector and Write_Sector take care of geometry translation for us,
-     * then call the absolute versions. So, we override the absolute versions only */
-    Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override {
-        unsigned char buffer[2048];
-
-	if (!src_drive)
-            return Int13Status::ControllerFailure;
-        if (!src_drive->ReadSectorsHost(buffer,false,cdrom_sector_offset+(sectnum>>2)/*512 byte/sector to 2048 byte/sector conversion*/,1))
-            return Int13Status::ControllerFailure;
-
-        if ((sectnum & 3) * 512 + 512 > sizeof(buffer)) return Int13Status::SectorNotFound;
-        memcpy(data,buffer+((sectnum&3)*512),512);
-        return Int13Status::NoError;
-    }
-    Int13Status Write_AbsoluteSector(uint32_t sectnum,const void * data) override {
-        (void)sectnum;//UNUSED
-        (void)data;//UNUSED
-        return Int13Status::WriteProtected; /* fail, read only */
-    }
-    imageDiskElToritoFloppy(unsigned char new_CDROM_drive,unsigned long new_cdrom_sector_offset,unsigned char floppy_emu_type) : imageDisk((FILE *)NULL,NULL,0,false), CDROM_drive(new_CDROM_drive), cdrom_sector_offset(new_cdrom_sector_offset), floppy_type(floppy_emu_type) {
-        diskimg = NULL;
-        sector_size = 512;
-        class_id = ID_EL_TORITO_FLOPPY;
-
-        bool GetMSCDEXDrive(unsigned char drive_letter,CDROM_Interface **_cdrom);
-        GetMSCDEXDrive(CDROM_drive-'A',&src_drive);/*addref src_drive*/
-
-        if (floppy_emu_type == 1) { /* 1.2MB */
-            heads = 2;
-            cylinders = 80;
-            sectors = 15;
-        }
-        else if (floppy_emu_type == 2) { /* 1.44MB */
-            heads = 2;
-            cylinders = 80;
-            sectors = 18;
-        }
-        else if (floppy_emu_type == 3) { /* 2.88MB */
-            heads = 2;
-            cylinders = 80;
-            sectors = 36; /* FIXME: right? */
-        }
-        else {
-            heads = 2;
-            cylinders = 69;
-            sectors = 14;
-            LOG_MSG("BUG! unsupported floppy_emu_type in El Torito floppy object\n");
-        }
-
-        diskSizeK = ((uint64_t)heads * cylinders * sectors * sector_size) / 1024;
-        active = true;
-    }
-    virtual ~imageDiskElToritoFloppy() {
-        if (src_drive) {
-            src_drive->Release();
-            src_drive = NULL;
-        }
-    }
-
-    CDROM_Interface *src_drive=NULL;
-    unsigned char CDROM_drive;
-    unsigned long cdrom_sector_offset;
-    unsigned char floppy_type;
-/*
-    int class_id;
-
-    bool hardDrive;
-    bool active;
-    FILE *diskimg;
-    std::string diskname;
-    uint8_t floppytype;
-
-    uint32_t sector_size;
-    uint32_t heads,cylinders,sectors;
-    uint32_t reserved_cylinders;
-    uint64_t current_fpos; */
-};
-
 /* PC-98 IPL1 partition table entry.
  * Taken from GNU Parted source code.
  * Maximum 16 entries. */
@@ -637,52 +441,5 @@ extern unsigned char INT13_ElTorito_NoEmuDriveNumber;
 extern signed char INT13_ElTorito_IDEInterface;
 extern CDROM_Interface * INT13_ElTorito_cdrom;
 extern char INT13_ElTorito_NoEmuCDROMDrive;
-
-#if !defined(OSFREE)
-class imageDiskINT13Drive : public imageDisk {
-public:
-	Int13Status Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size=0) override;
-	Int13Status Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size=0) override;
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	void UpdateFloppyType(void) override;
-	void Set_Reserved_Cylinders(Bitu resCyl) override;
-	uint32_t Get_Reserved_Cylinders() override;
-	void Set_Geometry(uint32_t setHeads, uint32_t setCyl, uint32_t setSect, uint32_t setSectSize) override;
-	void Get_Geometry(uint32_t * getHeads, uint32_t *getCyl, uint32_t *getSect, uint32_t *getSectSize) override;
-	uint8_t GetBiosType(void) override;
-	uint32_t getSectSize(void) override;
-	bool detectDiskChange(void) override;
-
-	imageDiskINT13Drive(imageDisk *sdisk);
-	virtual ~imageDiskINT13Drive();
-
-	uint8_t bios_disk = 0;
-	bool enable_int13 = false;
-	imageDisk* subdisk = NULL;
-	bool busy = false;
-};
-#endif
-
-#if !defined(OSFREE)
-class imageDiskMSDOSBlockDevice : public imageDisk {
-public:
-	Int13Status Read_AbsoluteSector(uint32_t sectnum, void * data) override;
-	Int13Status Write_AbsoluteSector(uint32_t sectnum, const void * data) override;
-
-	bool detectDiskChange(void) override;
-
-	imageDiskMSDOSBlockDevice();
-	virtual ~imageDiskMSDOSBlockDevice();
-
-	uint8_t media_dpb = 0;
-	uint8_t unit_code = 0;
-	uint16_t devseg = 0;
-	uint16_t attr = 0;
-	PhysPt devhdr = 0;
-	PhysPt bpbptr = 0;
-};
-#endif
 
 #endif
