+++ src/dos/dos_programs.cpp
@@ -45,7 +45,13 @@
 #include "cdrom.h"
 #include "builtin.h"
 #include "bios_disk.h"
+#include "imagedisk_d88.h"
+#include "imagedisk_eltorito.h"
+#include "imagedisk_emptydrive.h"
+#include "imagedisk_int13.h"
+#include "imagedisk_nfd.h"
 #include "imagedisk_teledisk.h"
+#include "imagedisk_vfd.h"
 #include "dos_system.h"
 #include "dos_inc.h"
 #include "bios.h"
