+++ src/dos/drive_fat.cpp
@@ -53,7 +53,12 @@
 #include "cross.h"
 #include "bios.h"
 #include "bios_disk.h"
+#include "imagedisk_d88.h"
+#include "imagedisk_eltorito.h"
+#include "imagedisk_int13.h"
+#include "imagedisk_nfd.h"
 #include "imagedisk_teledisk.h"
+#include "imagedisk_vfd.h"
 #include "qcow2_disk.h"
 #include "bitop.h"
 #include "callback.h"
