+++ src/ints/Makefile.am
@@ -5,4 +5,5 @@ libints_a_SOURCES = mouse.cpp xms.cpp xms.h ems.cpp int_dosv.cpp \
                     int10.cpp int10.h int10_char.cpp int10_memory.cpp int10_misc.cpp int10_modes.cpp \
                     int10_vesa.cpp int10_pal.cpp int10_put_pixel.cpp int10_video_state.cpp int10_vptable.cpp \
                     bios.cpp bios_disk.cpp bios_vhd.cpp bios_keyboard.cpp qcow2_disk.cpp bios_memdisk.cpp pc98_lio.cpp \
-                    imagedisk_teledisk.cpp
+                    imagedisk_d88.cpp imagedisk_emptydrive.cpp imagedisk_int13.cpp imagedisk_msdosblockdev.cpp \
+                    imagedisk_nfd.cpp imagedisk_teledisk.cpp imagedisk_vfd.cpp
