+++ src/ints/bios_disk.cpp
@@ -301,43 +301,6 @@ typedef struct {
 } FDIHDR;                       // =0x20
 
 typedef struct {
-	char	sig[16];            // +0x000
-	char	comment[0x100];     // +0x010
-	UINT8	headersize[4];      // +0x110
-    uint8_t prot;               // +0x114
-    uint8_t nhead;              // +0x115
-    uint8_t _unknown_[10];      // +0x116
-} NFDHDR;                       // =0x120
-
-typedef struct {
-	char	sig[16];            // +0x000
-	char	comment[0x100];     // +0x010
-	UINT8	headersize[4];      // +0x110
-    uint8_t prot;               // +0x114
-    uint8_t nhead;              // +0x115
-    uint8_t _unknown_[10];      // +0x116
-    uint32_t trackheads[164];   // +0x120
-    uint32_t addinfo;           // +0x3b0
-    uint8_t _unknown2_[12];     // +0x3b4
-} NFDHDRR1;                     // =0x3c0
-
-typedef struct {
-    uint8_t log_cyl;            // +0x0
-    uint8_t log_head;           // +0x1
-    uint8_t log_rec;            // +0x2
-    uint8_t sec_len_pow2;       // +0x3         sz = 128 << len_pow2
-    uint8_t flMFM;              // +0x4
-    uint8_t flDDAM;             // +0x5
-    uint8_t byStatus;           // +0x6
-    uint8_t bySTS0;             // +0x7
-    uint8_t bySTS1;             // +0x8
-    uint8_t bySTS2;             // +0x9
-    uint8_t byRetry;            // +0xA
-    uint8_t byPDA;              // +0xB
-    uint8_t _unknown_[4];       // +0xC
-} NFDHDR_ENTRY;                 // =0x10
-
-typedef struct {
     char        szFileID[15];                 // 識別ID "T98HDDIMAGE.R0"
     char        Reserve1[1];                  // 予約
     char        szComment[0x100];             // イメージコメント(ASCIIz)
@@ -2703,1078 +2666,6 @@ void BIOS_SetupDisks(void) {
     swapping_requested = false;
 }
 
-// VFD *.FDD floppy disk format support
-
-Int13Status imageDiskVFD::Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size) {
-    const vfdentry *ent;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-//    LOG_MSG("VFD read sector: CHS %u/%u/%u sz=%u",cylinder,head,sector,req_sector_size);
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    if (ent->hasSectorData()) {
-        fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-        if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-        if (fread(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-        return Int13Status::NoError;
-    }
-    else if (ent->hasFill()) {
-        memset(data,ent->fillbyte,req_sector_size);
-        return Int13Status::NoError;
-    }
-
-    return Int13Status::SectorNotFound;
-}
-
-Int13Status imageDiskVFD::Read_AbsoluteSector(uint32_t sectnum, void * data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Read_Sector(h,c,s,data);
-}
-
-imageDiskVFD::vfdentry *imageDiskVFD::findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size) {
-    std::vector<imageDiskVFD::vfdentry>::iterator i = dents.begin();
-    unsigned char szb=0xFF;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-    if (req_sector_size != ~0U) {
-        unsigned int c = req_sector_size;
-        while (c >= 128U) {
-            c >>= 1U;
-            szb++;
-        }
-
-//        LOG_MSG("req=%u c=%u szb=%u",req_sector_size,c,szb);
-
-        if (szb > 8 || c != 64U)
-            return NULL;
-    }
-
-    while (i != dents.end()) {
-        const imageDiskVFD::vfdentry &ent = *i;
-
-        if (ent.head == head &&
-            ent.track == track &&
-            ent.sector == sector &&
-            (ent.sizebyte == szb || req_sector_size == ~0U))
-            return &(*i);
-
-        ++i;
-    }
-
-    return NULL;
-}
-
-Int13Status imageDiskVFD::Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size) {
-    unsigned long new_offset;
-    unsigned char tmp[12];
-    vfdentry *ent;
-
-//    LOG_MSG("VFD write sector: CHS %u/%u/%u",cylinder,head,sector);
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    if (ent->hasSectorData()) {
-        fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-        if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-        if (fwrite(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-        return Int13Status::NoError;
-    }
-    else if (ent->hasFill()) {
-        bool isfill = false;
-
-        /* well, is the data provided one character repeated?
-         * note the format cannot represent a fill byte of 0xFF */
-        if (((unsigned char*)data)[0] != 0xFF) {
-            unsigned int i=1;
-
-            do {
-                if (((unsigned char*)data)[i] == ((unsigned char*)data)[0]) {
-                    if ((++i) == req_sector_size) {
-                        isfill = true;
-                        break; // yes!
-                    }
-                }
-                else {
-                    break; // nope
-                }
-            } while (1);
-        }
-
-        if (ent->entry_offset == 0) return Int13Status::SectorNotFound;
-
-        if (isfill) {
-            fseek(diskimg,(long)ent->entry_offset,SEEK_SET);
-            if ((uint32_t)ftell(diskimg) != ent->entry_offset) return Int13Status::SeekFailed;
-            if (fread(tmp,12,1,diskimg) != 1) return Int13Status::ControllerFailure;
-
-            tmp[0x04] = ((unsigned char*)data)[0]; // change the fill byte
-
-            LOG_MSG("VFD write: 'fill' sector changing fill byte to 0x%x",tmp[0x04]);
-
-            fseek(diskimg,(long)ent->entry_offset,SEEK_SET);
-            if ((uint32_t)ftell(diskimg) != ent->entry_offset) return Int13Status::SeekFailed;
-            if (fwrite(tmp,12,1,diskimg) != 1) return Int13Status::ControllerFailure;
-        }
-        else {
-            fseek(diskimg,0,SEEK_END);
-            new_offset = (unsigned long)ftell(diskimg);
-
-            /* we have to change it from a fill sector to an actual sector */
-            LOG_MSG("VFD write: changing 'fill' sector to one with data (data at %lu)",(unsigned long)new_offset);
-
-            fseek(diskimg,(long)ent->entry_offset,SEEK_SET);
-            if ((uint32_t)ftell(diskimg) != ent->entry_offset) return Int13Status::SeekFailed;
-            if (fread(tmp,12,1,diskimg) != 1) return Int13Status::ControllerFailure;
-
-            tmp[0x00] = ent->track;
-            tmp[0x01] = ent->head;
-            tmp[0x02] = ent->sector;
-            tmp[0x03] = ent->sizebyte;
-            tmp[0x04] = 0xFF; // no longer a fill byte
-            tmp[0x05] = 0x00; // TODO ??
-            tmp[0x06] = 0x00; // TODO ??
-            tmp[0x07] = 0x00; // TODO ??
-            *((uint32_t*)(tmp+8)) = new_offset;
-            ent->fillbyte = 0xFF;
-            ent->data_offset = (uint32_t)new_offset;
-
-            fseek(diskimg,(long)ent->entry_offset,SEEK_SET);
-            if ((uint32_t)ftell(diskimg) != ent->entry_offset) return Int13Status::SeekFailed;
-            if (fwrite(tmp,12,1,diskimg) != 1) return Int13Status::ControllerFailure;
-
-            fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-            if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-            if (fwrite(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-        }
-
-        return Int13Status::NoError;
-    }
-
-    return Int13Status::SectorNotFound;
-}
-
-Int13Status imageDiskVFD::Write_AbsoluteSector(uint32_t sectnum,const void *data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Write_Sector(h,c,s,data);
-}
-
-imageDiskVFD::imageDiskVFD(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk) : imageDisk(ID_VFD) {
-    (void)isHardDisk;//UNUSED
-    unsigned char tmp[16];
-
-    heads = 1;
-    cylinders = 0;
-    image_base = 0;
-    sectors = 0;
-    active = false;
-    sector_size = 0;
-    reserved_cylinders = 0;
-    diskSizeK = imgSizeK;
-    diskimg = imgFile;
-
-    if (imgName != NULL)
-        diskname = imgName;
-
-    // NOTES:
-    // 
-    //  +0x000: "VFD1.00"
-    //  +0x0DC: array of 12-byte entries each describing a sector
-    //
-    //  Each entry:
-    //  +0x0: track
-    //  +0x1: head
-    //  +0x2: sector
-    //  +0x3: sector size (128 << this byte)
-    //  +0x4: fill byte, or 0xFF
-    //  +0x5: unknown
-    //  +0x6: unknown
-    //  +0x7: unknown
-    //  +0x8: absolute data offset (32-bit integer) or 0xFFFFFFFF if the entire sector is that fill byte
-    fseek(diskimg,0,SEEK_SET);
-    memset(tmp,0,8);
-    size_t readResult = fread(tmp,1,8,diskimg);
-    if (readResult != 8) {
-            LOG(LOG_IO, LOG_ERROR) ("Reading error in imageDiskVFD constructor\n");
-            return;
-    }
-
-    if (!memcmp(tmp,"VFD1.",5)) {
-        uint32_t stop_at = 0xC3FC;
-        unsigned long entof;
-
-        // load table.
-        // we have to determine as we go where to stop reading.
-        // the source of info I read assumes the whole header (and table)
-        // is 0xC3FC bytes. I'm not inclined to assume that, so we go by
-        // that OR the first sector offset whichever is smaller.
-        // the table seems to trail off into a long series of 0xFF at the end.
-        fseek(diskimg,0xDC,SEEK_SET);
-        while ((entof=((unsigned long)ftell(diskimg)+12ul)) <= stop_at) {
-            memset(tmp,0xFF,12);
-            readResult = fread(tmp,12,1,diskimg);
-            if (readResult != 1) {
-                LOG(LOG_IO, LOG_ERROR) ("Reading error in imageDiskVFD constructor\n");
-                return;
-            }
-
-            if (!memcmp(tmp,"\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF\xFF",12))
-                continue;
-            if (!memcmp(tmp,"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",12))
-                continue;
-
-            struct vfdentry v;
-
-            v.track = tmp[0];
-            v.head = tmp[1];
-            v.sector = tmp[2];
-            v.sizebyte = tmp[3];
-            v.fillbyte = tmp[4];
-            v.data_offset = *((uint32_t*)(tmp+8));
-            v.entry_offset = (uint32_t)entof;
-
-            // maybe the table can end sooner than 0xC3FC?
-            // if we see sectors appear at an offset lower than our stop_at point
-            // then adjust the stop_at point. assume the table cannot mix with
-            // sector data.
-            if (v.hasSectorData()) {
-                if (stop_at > v.data_offset)
-                    stop_at = v.data_offset;
-            }
-
-            dents.push_back(v);
-
-            LOG_MSG("VFD entry: track=%u head=%u sector=%u size=%u fill=0x%2X has_data=%u has_fill=%u entoff=%lu dataoff=%lu",
-                v.track,
-                v.head,
-                v.sector,
-                v.getSectorSize(),
-                v.fillbyte,
-                v.hasSectorData(),
-                v.hasFill(),
-                (unsigned long)v.entry_offset,
-                (unsigned long)v.data_offset);
-        }
-
-        if (!dents.empty()) {
-            /* okay, now to figure out what the geometry of the disk is.
-             * we cannot just work from an "absolute" disk image model
-             * because there's no VFD header to just say what the geometry is.
-             * Like the IBM PC BIOS, we have to look at the disk and figure out
-             * which geometry to apply to it, even if the FDD format allows
-             * sectors on other tracks to have wild out of range sector, track,
-             * and head numbers or odd sized sectors.
-             *
-             * First, determine sector size according to the boot sector. */
-            const vfdentry *ent;
-
-            ent = findSector(/*head*/0,/*track*/0,/*sector*/1,~0U);
-            if (ent != NULL) {
-                if (ent->sizebyte <= 3) /* x <= 1024 */
-                    sector_size = ent->getSectorSize();
-            }
-
-            /* oh yeah right, sure.
-             * I suppose you're one of those FDD images where the sector size is 128 bytes/sector
-             * in the boot sector and the rest is 256 bytes/sector elsewhere. I have no idea why
-             * but quite a few FDD images have this arrangement. */
-            if (sector_size != 0 && sector_size < 512) {
-                ent = findSector(/*head*/0,/*track*/1,/*sector*/1,~0U);
-                if (ent != NULL) {
-                    if (ent->sizebyte <= 3) { /* x <= 1024 */
-                        unsigned int nsz = ent->getSectorSize();
-                        if (sector_size != nsz)
-                            LOG_MSG("VFD warning: sector size changes between track 0 and 1");
-                        if (sector_size < nsz)
-                            sector_size = nsz;
-                    }
-                }
-            }
-
-            uint8_t i;
-            if (sector_size != 0) {
-                i=0;
-                while (DiskGeometryList[i].ksize != 0) {
-                    const diskGeo &diskent = DiskGeometryList[i];
-
-                    if (diskent.bytespersect == sector_size) {
-                        ent = findSector(0,0,diskent.secttrack);
-                        if (ent != NULL) {
-                            LOG_MSG("VFD disk probe: %u/%u/%u exists",0,0,diskent.secttrack);
-                            if (sectors < diskent.secttrack)
-                                sectors = diskent.secttrack;
-                        }
-                    }
-
-                    i++;
-                }
-            }
-
-            if (sector_size != 0 && sectors != 0) {
-                i=0;
-                while (DiskGeometryList[i].ksize != 0) {
-                    const diskGeo &diskent = DiskGeometryList[i];
-
-                    if (diskent.bytespersect == sector_size && diskent.secttrack >= sectors) {
-                        ent = findSector(0,diskent.cylcount-1,sectors);
-                        if (ent != NULL) {
-                            LOG_MSG("VFD disk probe: %u/%u/%u exists",0,diskent.cylcount-1,sectors);
-                            if (cylinders < diskent.cylcount)
-                                cylinders = diskent.cylcount;
-                        }
-                    }
-
-                    i++;
-                }
-            }
-
-            if (sector_size != 0 && sectors != 0 && cylinders != 0) {
-                ent = findSector(1,0,sectors);
-                if (ent != NULL) {
-                    LOG_MSG("VFD disk probe: %u/%u/%u exists",1,0,sectors);
-                    heads = 2;
-                }
-            }
-
-            // TODO: drive_fat.cpp should use an extension to this API to allow changing the sectors/track
-            //       according to what it reads from the MS-DOS BIOS parameter block, just like real MS-DOS.
-            //       This would allow better representation of strange disk formats such as the "extended"
-            //       floppy format that Microsoft used to use for Word 95 and Windows 95 install floppies.
-
-            LOG_MSG("VFD geometry detection: C/H/S %u/%u/%u %u bytes/sector",
-                cylinders, heads, sectors, sector_size);
-
-            bool founddisk = false;
-            if (sector_size != 0 && sectors != 0 && cylinders != 0 && heads != 0)
-                founddisk = true;
-
-            if(!founddisk) {
-                active = false;
-            } else {
-                incrementFDD();
-                updateFloppyDPT();
-            }
-        }
-    }
-}
-
-imageDiskVFD::~imageDiskVFD() {
-    if(diskimg != NULL) {
-        fclose(diskimg);
-        diskimg=NULL; 
-    }
-}
-
-// D88 *.D88 floppy disk format support
-
-enum {
-    D88_TRACKMAX        = 164,
-    D88_HEADERSIZE      = 0x20 + (D88_TRACKMAX * 4)
-};
-
-#pragma pack(push,1)
-typedef struct D88HEAD {
-    char            fd_name[17];                // +0x00 Disk Name
-    unsigned char   reserved1[9];               // +0x11 Reserved
-    unsigned char   protect;                    // +0x1A Write Protect bit:4
-    unsigned char   fd_type;                    // +0x1B Disk Format
-    uint32_t        fd_size;                    // +0x1C Disk Size
-    uint32_t        trackp[D88_TRACKMAX];       // +0x20 <array of DWORDs>     164 x 4 = 656 = 0x290
-} D88HEAD;                                      // =0x2B0 total
-
-typedef struct D88SEC {
-    unsigned char   c;                          // +0x00
-    unsigned char   h;                          // +0x01
-    unsigned char   r;                          // +0x02
-    unsigned char   n;                          // +0x03
-    uint16_t        sectors;                    // +0x04 Sector Count
-    unsigned char   mfm_flg;                    // +0x06 sides
-    unsigned char   del_flg;                    // +0x07 DELETED DATA
-    unsigned char   stat;                       // +0x08 STATUS (FDC ret)
-    unsigned char   seektime;                   // +0x09 Seek Time
-    unsigned char   reserved[3];                // +0x0A Reserved
-    unsigned char   rpm_flg;                    // +0x0D rpm          0:1.2  1:1.44
-    uint16_t        size;                       // +0x0E Sector Size
-                                                // <sector contents follow>
-} D88SEC;                                       // =0x10 total
-#pragma pack(pop)
-
-Int13Status imageDiskD88::Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size) {
-    const vfdentry *ent;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-//    LOG_MSG("D88 read sector: CHS %u/%u/%u sz=%u",cylinder,head,sector,req_sector_size);
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-    if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-    if (fread(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-    return Int13Status::NoError;
-}
-
-Int13Status imageDiskD88::Read_AbsoluteSector(uint32_t sectnum, void * data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Read_Sector(h,c,s,data);
-}
-
-imageDiskD88::vfdentry *imageDiskD88::findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size) {
-    if ((size_t)track >= dents.size())
-        return NULL;
-
-    std::vector<imageDiskD88::vfdentry>::iterator i = dents.begin();
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-    while (i != dents.end()) {
-        const imageDiskD88::vfdentry &ent = *i;
-
-        if (ent.head == head &&
-            ent.track == track &&
-            ent.sector == sector &&
-            (ent.sector_size == req_sector_size || req_sector_size == ~0U))
-            return &(*i);
-
-        ++i;
-    }
-
-    return NULL;
-}
-
-Int13Status imageDiskD88::Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size) {
-    const vfdentry *ent;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-//    LOG_MSG("D88 read sector: CHS %u/%u/%u sz=%u",cylinder,head,sector,req_sector_size);
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-    if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-    if (fwrite(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-    return Int13Status::NoError;
-}
-
-Int13Status imageDiskD88::Write_AbsoluteSector(uint32_t sectnum,const void *data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Write_Sector(h,c,s,data);
-}
-
-imageDiskD88::imageDiskD88(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk) : imageDisk(ID_D88) {
-    (void)isHardDisk;//UNUSED
-    D88HEAD head;
-
-    fd_type_major = DISKTYPE_2D;
-    fd_type_minor = 0;
-
-    assert(sizeof(D88HEAD) == 0x2B0);
-    assert(sizeof(D88SEC) == 0x10);
-
-    heads = 0;
-    cylinders = 0;
-    image_base = 0;
-    sectors = 0;
-    active = false;
-    sector_size = 0;
-    reserved_cylinders = 0;
-    diskSizeK = imgSizeK;
-    diskimg = imgFile;
-
-    if (imgName != NULL)
-        diskname = imgName;
-
-    // NOTES:
-    // 
-    //  +0x000: D88 header
-    //  +0x020: Offset of D88 tracks, per track
-    //  +0x2B0: <begin data>
-    //
-    // Track offsets are sequential, always
-    //
-    // Each track is an array of:
-    //
-    //  ENTRY:
-    //   <D88 sector head>
-    //   <sector contents>
-    //
-    // Array of ENTRY from offset until next track
-    fseek(diskimg,0,SEEK_END);
-    off_t fsz = ftell(diskimg);
-
-    fseek(diskimg,0,SEEK_SET);
-    if (fread(&head,sizeof(head),1,diskimg) != 1) return;
-
-    // validate fd_size
-    if (host_readd((ConstHostPt)(&head.fd_size)) > (uint32_t)fsz) return;
-
-    fd_type_major = head.fd_type >> 4U;
-    fd_type_minor = head.fd_type & 0xFU;
-
-    // validate that none of the track offsets extend past the file
-    {
-        for (unsigned int i=0;i < D88_TRACKMAX;i++) {
-            uint32_t trackoff = host_readd((ConstHostPt)(&head.trackp[i]));
-
-            if (trackoff == 0) continue;
-
-            if ((trackoff + 16U) > (uint32_t)fsz) {
-                LOG_MSG("D88: track starts past end of file");
-                return;
-            }
-        }
-    }
-
-    // read each track
-    for (unsigned int track=0;track < D88_TRACKMAX;track++) {
-        uint32_t trackoff = host_readd((ConstHostPt)(&head.trackp[track]));
-
-        if (trackoff != 0) {
-            fseek(diskimg, (long)trackoff, SEEK_SET);
-            if ((off_t)ftell(diskimg) != (off_t)trackoff) continue;
-
-            D88SEC s;
-            unsigned int count = 0;
-
-            do {
-                if (fread(&s,sizeof(s),1,diskimg) != 1) break;
-
-                uint16_t sector_count = host_readw((ConstHostPt)(&s.sectors));
-                uint16_t sector_size = host_readw((ConstHostPt)(&s.size));
-
-                if (sector_count == 0U || sector_size < 128U) break;
-                if (sector_count > 128U || sector_size > 16384U) break;
-                if (s.n > 8U) s.n = 8U;
-
-                vfdentry vent;
-                vent.sector_size = 128 << s.n;
-                vent.data_offset = (uint32_t)ftell(diskimg);
-                vent.entry_offset = vent.data_offset - (uint32_t)16;
-                vent.track = s.c;
-                vent.head = s.h;
-                vent.sector = s.r;
-
-                LOG_MSG("D88: trackindex=%u C/H/S/sz=%u/%u/%u/%u data-at=0x%lx",
-                    track,vent.track,vent.head,vent.sector,vent.sector_size,(unsigned long)vent.data_offset);
-
-                dents.push_back(vent);
-                if ((++count) >= sector_count) break;
-
-                fseek(diskimg, (long)sector_size, SEEK_CUR);
-            } while (1);
-        }
-    }
-
-    if (!dents.empty()) {
-        /* okay, now to figure out what the geometry of the disk is.
-         * we cannot just work from an "absolute" disk image model
-         * because there's no D88 header to just say what the geometry is.
-         * Like the IBM PC BIOS, we have to look at the disk and figure out
-         * which geometry to apply to it, even if the FDD format allows
-         * sectors on other tracks to have wild out of range sector, track,
-         * and head numbers or odd sized sectors.
-         *
-         * First, determine sector size according to the boot sector. */
-        bool founddisk = false;
-        const vfdentry *ent;
-
-        ent = findSector(/*head*/0,/*track*/0,/*sector*/1,~0U);
-        if (ent != NULL) {
-            if (ent->getSectorSize() <= 1024) /* x <= 1024 */
-                sector_size = ent->getSectorSize();
-        }
-
-        /* oh yeah right, sure.
-         * I suppose you're one of those FDD images where the sector size is 128 bytes/sector
-         * in the boot sector and the rest is 256 bytes/sector elsewhere. I have no idea why
-         * but quite a few FDD images have this arrangement. */
-        if (sector_size != 0 && sector_size < 512) {
-            ent = findSector(/*head*/0,/*track*/1,/*sector*/1,~0U);
-            if (ent != NULL) {
-                if (ent->getSectorSize() <= 1024) { /* x <= 1024 */
-                    unsigned int nsz = ent->getSectorSize();
-                    if (sector_size != nsz)
-                        LOG_MSG("D88 warning: sector size changes between track 0 and 1");
-                    if (sector_size < nsz)
-                        sector_size = nsz;
-                }
-            }
-        }
-
-        if (sector_size != 0) {
-            unsigned int i = 0;
-            while (DiskGeometryList[i].ksize != 0) {
-                const diskGeo &diskent = DiskGeometryList[i];
-
-                if (diskent.bytespersect == sector_size) {
-                    ent = findSector(0,0,diskent.secttrack);
-                    if (ent != NULL) {
-                        LOG_MSG("D88 disk probe: %u/%u/%u exists",0,0,diskent.secttrack);
-                        if (sectors < diskent.secttrack)
-                            sectors = diskent.secttrack;
-                    }
-                }
-
-                i++;
-            }
-        }
-
-        if (sector_size != 0 && sectors != 0) {
-            unsigned int i = 0;
-            while (DiskGeometryList[i].ksize != 0) {
-                const diskGeo &diskent = DiskGeometryList[i];
-
-                if (diskent.bytespersect == sector_size && diskent.secttrack >= sectors) {
-                    ent = findSector(0,diskent.cylcount-1,sectors);
-                    if (ent != NULL) {
-                        LOG_MSG("D88 disk probe: %u/%u/%u exists",0,diskent.cylcount-1,sectors);
-                        if (cylinders < diskent.cylcount)
-                            cylinders = diskent.cylcount;
-                    }
-                }
-
-                i++;
-            }
-        }
-
-        if (sector_size != 0 && sectors != 0 && cylinders != 0) {
-            ent = findSector(1,0,sectors);
-            if (ent != NULL) {
-                LOG_MSG("D88 disk probe: %u/%u/%u exists",1,0,sectors);
-                heads = 2;
-            }
-        }
-
-        // TODO: drive_fat.cpp should use an extension to this API to allow changing the sectors/track
-        //       according to what it reads from the MS-DOS BIOS parameter block, just like real MS-DOS.
-        //       This would allow better representation of strange disk formats such as the "extended"
-        //       floppy format that Microsoft used to use for Word 95 and Windows 95 install floppies.
-
-        LOG_MSG("D88 geometry detection: C/H/S %u/%u/%u %u bytes/sector",
-                cylinders, heads, sectors, sector_size);
-
-        if (sector_size != 0 && sectors != 0 && cylinders != 0 && heads != 0)
-            founddisk = true;
-
-        if(!founddisk) {
-            active = false;
-        } else {
-            incrementFDD();
-            updateFloppyDPT();
-        }
-    }
-}
-
-imageDiskD88::~imageDiskD88() {
-    if(diskimg != NULL) {
-        fclose(diskimg);
-        diskimg=NULL; 
-    }
-}
-
-/*--------------------------------*/
-
-Int13Status imageDiskNFD::Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size) {
-    const vfdentry *ent;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-//    LOG_MSG("NFD read sector: CHS %u/%u/%u sz=%u",cylinder,head,sector,req_sector_size);
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-    if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-    if (fread(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-    return Int13Status::NoError;
-}
-
-Int13Status imageDiskNFD::Read_AbsoluteSector(uint32_t sectnum, void * data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Read_Sector(h,c,s,data);
-}
-
-imageDiskNFD::vfdentry *imageDiskNFD::findSector(uint8_t head,uint8_t track,uint8_t sector/*TODO: physical head?*/,unsigned int req_sector_size) {
-    if ((size_t)track >= dents.size())
-        return NULL;
-
-    std::vector<imageDiskNFD::vfdentry>::iterator i = dents.begin();
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-    while (i != dents.end()) {
-        const imageDiskNFD::vfdentry &ent = *i;
-
-        if (ent.head == head &&
-            ent.track == track &&
-            ent.sector == sector &&
-            (ent.sector_size == req_sector_size || req_sector_size == ~0U))
-            return &(*i);
-
-        ++i;
-    }
-
-    return NULL;
-}
-
-Int13Status imageDiskNFD::Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size) {
-    const vfdentry *ent;
-
-    if (req_sector_size == 0)
-        req_sector_size = sector_size;
-
-//    LOG_MSG("NFD read sector: CHS %u/%u/%u sz=%u",cylinder,head,sector,req_sector_size);
-
-    ent = findSector(head,cylinder,sector,req_sector_size);
-    if (ent == NULL) return Int13Status::SectorNotFound;
-    if (ent->getSectorSize() != req_sector_size) return Int13Status::SectorNotFound;
-
-    fseek(diskimg,(long)ent->data_offset,SEEK_SET);
-    if ((uint32_t)ftell(diskimg) != ent->data_offset) return Int13Status::SeekFailed;
-    if (fwrite(data,req_sector_size,1,diskimg) != 1) return Int13Status::ControllerFailure;
-    return Int13Status::NoError;
-}
-
-Int13Status imageDiskNFD::Write_AbsoluteSector(uint32_t sectnum,const void *data) {
-    unsigned int c,h,s;
-
-    if (sectors == 0 || heads == 0)
-        return Int13Status::SectorNotFound;
-
-    s = (sectnum % sectors) + 1;
-    h = (sectnum / sectors) % heads;
-    c = (sectnum / sectors / heads);
-    return Write_Sector(h,c,s,data);
-}
-
-imageDiskNFD::imageDiskNFD(FILE *imgFile, const char *imgName, uint32_t imgSizeK, bool isHardDisk, unsigned int revision) : imageDisk(ID_NFD) {
-    (void)isHardDisk;//UNUSED
-    union {
-        NFDHDR head;
-        NFDHDRR1 headr1;
-    }; // these occupy the same location of memory
-
-    assert(sizeof(NFDHDR) == 0x120);
-    assert(sizeof(NFDHDRR1) == 0x3C0);
-    assert(sizeof(NFDHDR_ENTRY) == 0x10);
-
-    heads = 0;
-    cylinders = 0;
-    image_base = 0;
-    sectors = 0;
-    active = false;
-    sector_size = 0;
-    reserved_cylinders = 0;
-    diskSizeK = imgSizeK;
-    diskimg = imgFile;
-
-    if (imgName != NULL)
-        diskname = imgName;
-
-    // NOTES:
-    // 
-    //  +0x000: NFD header
-    //  +0x020: Offset of NFD tracks, per track
-    //  +0x2B0: <begin data>
-    //
-    // Track offsets are sequential, always
-    //
-    // Each track is an array of:
-    //
-    //  ENTRY:
-    //   <NFD sector head>
-    //   <sector contents>
-    //
-    // Array of ENTRY from offset until next track
-    fseek(diskimg,0,SEEK_END);
-    off_t fsz = ftell(diskimg);
-
-    fseek(diskimg,0,SEEK_SET);
-    if (revision == 0) {
-        if (fread(&head,sizeof(head),1,diskimg) != 1) return;
-    }
-    else if (revision == 1) {
-        if (fread(&headr1,sizeof(headr1),1,diskimg) != 1) return;
-    }
-    else {
-        abort();
-    }
-
-    // validate fd_size
-    if (host_readd((ConstHostPt)(&head.headersize)) < sizeof(head)) return;
-    if (host_readd((ConstHostPt)(&head.headersize)) > (uint32_t)fsz) return;
-
-    unsigned int data_offset = host_readd((ConstHostPt)(&head.headersize));
-
-    std::vector< std::pair<uint32_t,NFDHDR_ENTRY> > seclist;
-
-    if (revision == 0) {
-        unsigned int secents = (host_readd((ConstHostPt)(&head.headersize)) - sizeof(head)) / sizeof(NFDHDR_ENTRY);
-        if (secents == 0) return;
-        secents--;
-        if (secents == 0) return;
-
-        for (unsigned int i=0;i < secents;i++) {
-            uint32_t ofs = (uint32_t)ftell(diskimg);
-            NFDHDR_ENTRY e;
-
-            if (fread(&e,sizeof(e),1,diskimg) != 1) return;
-            seclist.push_back( std::pair<uint32_t,NFDHDR_ENTRY>(ofs,e) );
-
-            if (e.log_cyl == 0xFF || e.log_head == 0xFF || e.log_rec == 0xFF || e.sec_len_pow2 > 7)
-                continue;
-
-            LOG_MSG("NFD %u/%u: ofs=%lu data=%lu cyl=%u head=%u sec=%u len=%u",
-                    (unsigned int)i,
-                    (unsigned int)secents,
-                    (unsigned long)ofs,
-                    (unsigned long)data_offset,
-                    e.log_cyl,
-                    e.log_head,
-                    e.log_rec,
-                    128 << e.sec_len_pow2);
-
-            vfdentry vent;
-            vent.sector_size = 128 << e.sec_len_pow2;
-            vent.data_offset = (uint32_t)data_offset;
-            vent.entry_offset = ofs;
-            vent.track = e.log_cyl;
-            vent.head = e.log_head;
-            vent.sector = e.log_rec;
-            dents.push_back(vent);
-
-            data_offset += 128u << e.sec_len_pow2;
-            if (data_offset > (unsigned int)fsz) return;
-        }
-    }
-    else {
-        /* R1 has an array of offsets to where each tracks begins.
-         * The end of the track is an entry like 0x1A 0x00 0x00 0x00 0x00 0x00 0x00 .... */
-        /* The R1 images I have as reference always have offsets in ascending order. */
-        for (unsigned int ti=0;ti < 164;ti++) {
-            uint32_t trkoff = host_readd((ConstHostPt)(&headr1.trackheads[ti]));
-
-            if (trkoff == 0) break;
-
-            fseek(diskimg,(long)trkoff,SEEK_SET);
-            if ((off_t)ftell(diskimg) != (off_t)trkoff) return;
-
-            NFDHDR_ENTRY e;
-
-            // track id
-            if (fread(&e,sizeof(e),1,diskimg) != 1) return;
-            unsigned int sectors = host_readw((ConstHostPt)(&e) + 0);
-            unsigned int diagcount = host_readw((ConstHostPt)(&e) + 2);
-
-            LOG_MSG("NFD R1 track ent %u offset %lu sectors %u diag %u",ti,(unsigned long)trkoff,sectors,diagcount);
-
-            for (unsigned int s=0;s < sectors;s++) {
-                uint32_t ofs = (uint32_t)ftell(diskimg);
-
-                if (fread(&e,sizeof(e),1,diskimg) != 1) return;
-
-                LOG_MSG("NFD %u/%u: ofs=%lu data=%lu cyl=%u head=%u sec=%u len=%u rep=%u",
-                        s,
-                        sectors,
-                        (unsigned long)ofs,
-                        (unsigned long)data_offset,
-                        e.log_cyl,
-                        e.log_head,
-                        e.log_rec,
-                        128 << e.sec_len_pow2,
-                        e.byRetry);
-
-                vfdentry vent;
-                vent.sector_size = 128 << e.sec_len_pow2;
-                vent.data_offset = (uint32_t)data_offset;
-                vent.entry_offset = ofs;
-                vent.track = e.log_cyl;
-                vent.head = e.log_head;
-                vent.sector = e.log_rec;
-                dents.push_back(vent);
-
-                data_offset += 128u << e.sec_len_pow2;
-                if (data_offset > (unsigned int)fsz) return;
-            }
-
-            for (unsigned int d=0;d < diagcount;d++) {
-                if (fread(&e,sizeof(e),1,diskimg) != 1) return;
-
-                unsigned int retry = e.byRetry;
-                unsigned int len = host_readd((ConstHostPt)(&e) + 10);
-
-                LOG_MSG("NFD diag %u/%u: retry=%u len=%u data=%lu",d,diagcount,retry,len,(unsigned long)data_offset);
-
-                data_offset += (1+retry) * len;
-            }
-        }
-    }
-
-    if (!dents.empty()) {
-        /* okay, now to figure out what the geometry of the disk is.
-         * we cannot just work from an "absolute" disk image model
-         * because there's no NFD header to just say what the geometry is.
-         * Like the IBM PC BIOS, we have to look at the disk and figure out
-         * which geometry to apply to it, even if the FDD format allows
-         * sectors on other tracks to have wild out of range sector, track,
-         * and head numbers or odd sized sectors.
-         *
-         * First, determine sector size according to the boot sector. */
-        bool founddisk = false;
-        const vfdentry *ent;
-
-        ent = findSector(/*head*/0,/*track*/0,/*sector*/1,~0U);
-        if (ent != NULL) {
-            if (ent->getSectorSize() <= 1024) /* x <= 1024 */
-                sector_size = ent->getSectorSize();
-        }
-
-        /* oh yeah right, sure.
-         * I suppose you're one of those FDD images where the sector size is 128 bytes/sector
-         * in the boot sector and the rest is 256 bytes/sector elsewhere. I have no idea why
-         * but quite a few FDD images have this arrangement. */
-        if (sector_size != 0 && sector_size < 512) {
-            ent = findSector(/*head*/0,/*track*/1,/*sector*/1,~0U);
-            if (ent != NULL) {
-                if (ent->getSectorSize() <= 1024) { /* x <= 1024 */
-                    unsigned int nsz = ent->getSectorSize();
-                    if (sector_size != nsz)
-                        LOG_MSG("NFD warning: sector size changes between track 0 and 1");
-                    if (sector_size < nsz)
-                        sector_size = nsz;
-                }
-            }
-        }
-
-        if (sector_size != 0) {
-            unsigned int i = 0;
-            while (DiskGeometryList[i].ksize != 0) {
-                const diskGeo &diskent = DiskGeometryList[i];
-
-                if (diskent.bytespersect == sector_size) {
-                    ent = findSector(0,0,diskent.secttrack);
-                    if (ent != NULL) {
-                        LOG_MSG("NFD disk probe: %u/%u/%u exists",0,0,diskent.secttrack);
-                        if (sectors < diskent.secttrack)
-                            sectors = diskent.secttrack;
-                    }
-                }
-
-                i++;
-            }
-        }
-
-        if (sector_size != 0 && sectors != 0) {
-            unsigned int i = 0;
-            while (DiskGeometryList[i].ksize != 0) {
-                const diskGeo &diskent = DiskGeometryList[i];
-
-                if (diskent.bytespersect == sector_size && diskent.secttrack >= sectors) {
-                    ent = findSector(0,diskent.cylcount-1,sectors);
-                    if (ent != NULL) {
-                        LOG_MSG("NFD disk probe: %u/%u/%u exists",0,diskent.cylcount-1,sectors);
-                        if (cylinders < diskent.cylcount)
-                            cylinders = diskent.cylcount;
-                    }
-                }
-
-                i++;
-            }
-        }
-
-        if (sector_size != 0 && sectors != 0 && cylinders != 0) {
-            ent = findSector(1,0,sectors);
-            if (ent != NULL) {
-                LOG_MSG("NFD disk probe: %u/%u/%u exists",1,0,sectors);
-                heads = 2;
-            }
-        }
-
-        // TODO: drive_fat.cpp should use an extension to this API to allow changing the sectors/track
-        //       according to what it reads from the MS-DOS BIOS parameter block, just like real MS-DOS.
-        //       This would allow better representation of strange disk formats such as the "extended"
-        //       floppy format that Microsoft used to use for Word 95 and Windows 95 install floppies.
-
-        LOG_MSG("NFD geometry detection: C/H/S %u/%u/%u %u bytes/sector",
-                cylinders, heads, sectors, sector_size);
-
-        if (sector_size != 0 && sectors != 0 && cylinders != 0 && heads != 0)
-            founddisk = true;
-
-        if(!founddisk) {
-            active = false;
-        } else {
-            incrementFDD();
-            updateFloppyDPT();
-        }
-    }
-}
-
-imageDiskNFD::~imageDiskNFD() {
-    if(diskimg != NULL) {
-        fclose(diskimg);
-        diskimg=NULL; 
-    }
-}
-
 bool PartitionLoadMBR(std::vector<partTable::partentry_t> &parts,imageDisk *loadedDisk) {
 	partTable smbr;
 
@@ -3924,287 +2815,4 @@ void LogPrintPartitionTable(const std::vector<partTable::partentry_t> &parts) {
 			(unsigned long long)part.partSize);
 	}
 }
-
-
-Int13Status imageDiskEmptyDrive::Read_Sector(uint32_t /*head*/,uint32_t /*cylinder*/,uint32_t /*sector*/,void * /*data*/,unsigned int /*req_sector_size*/) {
-	return Int13Status::DriveNotReady;
-}
-
-Int13Status imageDiskEmptyDrive::Write_Sector(uint32_t /*head*/,uint32_t /*cylinder*/,uint32_t /*sector*/,const void * /*data*/,unsigned int /*req_sector_size*/) {
-	return Int13Status::DriveNotReady;
-}
-
-Int13Status imageDiskEmptyDrive::Read_AbsoluteSector(uint32_t /*sectnum*/, void * /*data*/) {
-	return Int13Status::DriveNotReady;
-}
-
-Int13Status imageDiskEmptyDrive::Write_AbsoluteSector(uint32_t /*sectnum*/, const void * /*data*/) {
-	return Int13Status::DriveNotReady;
-}
-
-imageDiskEmptyDrive::imageDiskEmptyDrive() : imageDisk(ID_EMPTY_DRIVE) {
-	active = true;
-	sector_size = 512;
-	heads = 2;
-	cylinders = 80;
-	sectors = 18;
-	diskSizeK = 1440;
-}
-
-imageDiskEmptyDrive::~imageDiskEmptyDrive() {
-}
-
-/////
-
-#if !defined(OSFREE)
-unsigned int INT13Xfer = 0;
-size_t INT13XferSize = 4096;
-#endif
-
-#if !defined(OSFREE)
-static void imageDiskCallINT13(void) {
-	unsigned int rv = CALLBACK_RealPointer(call_int13);
-	Bitu oldIF=GETFLAG(IF);
-	SETFLAGBIT(IF,true);
-	uint16_t oldcs=SegValue(cs);
-	uint32_t oldeip=reg_eip;
-	SegSet16(cs,rv>>16);
-	reg_eip=(rv&0xFFFF)+4+5;
-	DOSBOX_RunMachine();
-	reg_eip=oldeip;
-	SegSet16(cs,oldcs);
-	SETFLAGBIT(IF,oldIF);
-}
-#endif
-
-#if !defined(OSFREE)
-Int13Status imageDiskINT13Drive::Read_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,void * data,unsigned int req_sector_size) {
-	if (!enable_int13 || busy) return subdisk->Read_Sector(head,cylinder,sector,data,req_sector_size);
-
-	Int13Status ret = Int13Status::ResetFailed;
-	unsigned int retry = 3;
-
-	if (req_sector_size == 0) req_sector_size = sector_size;
-
-	//LOG_MSG("INT13 read C/H/S %u/%u/%u busy=%u",cylinder,head,sector,busy);
-
-	if (!busy && sector_size == req_sector_size && sector_size <= INT13XferSize) {
-		busy = true;
-
-		if (INT13Xfer == 0) INT13Xfer = DOS_GetMemory(INT13XferSize/16u,"INT 13 transfer buffer");
-
-		unsigned int s_eax = reg_eax;
-		unsigned int s_ebx = reg_ebx;
-		unsigned int s_ecx = reg_ecx;
-		unsigned int s_edx = reg_edx;
-		unsigned int s_esi = reg_esi;
-		unsigned int s_edi = reg_edi;
-		unsigned int s_esp = reg_esp;
-		unsigned int s_ebp = reg_ebp;
-		unsigned int s_es  = SegValue(es);
-		unsigned int s_fl  = reg_flags;
-
-again:
-		reg_eax = 0x200/*read command*/ | 1/*count*/;
-		reg_ebx = 0;
-		reg_ch = cylinder;
-		reg_cl = sector;
-		reg_dh = head;
-		reg_dl = bios_disk;
-		CPU_SetSegGeneral(es,INT13Xfer);
-
-		imageDiskCallINT13();
-
-		if (reg_flags & FLAG_CF) {
-			ret = static_cast<Int13Status>(reg_ah);
-			if (ret == Int13Status::NoError) ret = Int13Status::ResetFailed;
-
-			if (ret == Int13Status::DiskChanged) {
-				diskChangeFlag = true;
-				if (--retry > 0) goto again;
-			}
-		}
-		else {
-			ret = Int13Status::NoError;
-			MEM_BlockRead32(INT13Xfer<<4,data,sector_size);
-			data = (void*)((char*)data + sector_size);
-			if ((++sector) >= (sectors + 1)) {
-				assert(sector == (sectors + 1));
-				sector = 1;
-				if ((++head) >= heads) {
-					assert(head == heads);
-					head = 0;
-					cylinder++;
-				}
-			}
-		}
-
-		reg_eax = s_eax;
-		reg_ebx = s_ebx;
-		reg_ecx = s_ecx;
-		reg_edx = s_edx;
-		reg_esi = s_esi;
-		reg_edi = s_edi;
-		reg_esp = s_esp;
-		reg_ebp = s_ebp;
-		reg_flags = s_fl;
-		CPU_SetSegGeneral(es,s_es);
-
-		busy = false;
-	}
-
-	return ret;
-}
-#endif
-
-#if !defined(OSFREE)
-Int13Status imageDiskINT13Drive::Write_Sector(uint32_t head,uint32_t cylinder,uint32_t sector,const void * data,unsigned int req_sector_size) {
-	if (INT13Xfer == 0) INT13Xfer = DOS_GetMemory(INT13XferSize/16u,"INT 13 transfer buffer");
-
-	return subdisk->Write_Sector(head,cylinder,sector,data,req_sector_size);
-}
-#endif
-
-#if !defined(OSFREE)
-Int13Status imageDiskINT13Drive::Read_AbsoluteSector(uint32_t sectnum, void * data) {
-	unsigned int c,h,s;
-
-	if (sectors == 0 || heads == 0)
-		return Int13Status::SectorNotFound;
-
-	s = (sectnum % sectors) + 1;
-	h = (sectnum / sectors) % heads;
-	c = (sectnum / sectors / heads);
-	return Read_Sector(h,c,s,data);
-}
-#endif
-
-#if !defined(OSFREE)
-Int13Status imageDiskINT13Drive::Write_AbsoluteSector(uint32_t sectnum, const void * data) {
-	unsigned int c,h,s;
-
-	if (sectors == 0 || heads == 0)
-		return Int13Status::SectorNotFound;
-
-	s = (sectnum % sectors) + 1;
-	h = (sectnum / sectors) % heads;
-	c = (sectnum / sectors / heads);
-	return Write_Sector(h,c,s,data);
-}
-#endif
-
-#if !defined(OSFREE)
-void imageDiskINT13Drive::UpdateFloppyType(void) {
-	subdisk->UpdateFloppyType();
-}
-#endif
-
-#if !defined(OSFREE)
-void imageDiskINT13Drive::Set_Reserved_Cylinders(Bitu resCyl) {
-	subdisk->Set_Reserved_Cylinders(resCyl);
-}
-#endif
-
-#if !defined(OSFREE)
-uint32_t imageDiskINT13Drive::Get_Reserved_Cylinders() {
-	return subdisk->Get_Reserved_Cylinders();
-}
-#endif
-
-#if !defined(OSFREE)
-void imageDiskINT13Drive::Set_Geometry(uint32_t setHeads, uint32_t setCyl, uint32_t setSect, uint32_t setSectSize) {
-	heads = setHeads;
-	cylinders = setCyl;
-	sectors = setSect;
-	sector_size = setSectSize;
-	return subdisk->Set_Geometry(setHeads,setCyl,setSect,setSectSize);
-}
-#endif
-
-#if !defined(OSFREE)
-void imageDiskINT13Drive::Get_Geometry(uint32_t * getHeads, uint32_t *getCyl, uint32_t *getSect, uint32_t *getSectSize) {
-	return subdisk->Get_Geometry(getHeads,getCyl,getSect,getSectSize);
-}
-#endif
-
-#if !defined(OSFREE)
-uint8_t imageDiskINT13Drive::GetBiosType(void) {
-	return subdisk->GetBiosType();
-}
-#endif
-
-#if !defined(OSFREE)
-uint32_t imageDiskINT13Drive::getSectSize(void) {
-	return subdisk->getSectSize();
-}
-#endif
-
-#if !defined(OSFREE)
-bool imageDiskINT13Drive::detectDiskChange(void) {
-	if (enable_int13 && !busy) {
-		busy = true;
-
-		unsigned int s_eax = reg_eax;
-		unsigned int s_ebx = reg_ebx;
-		unsigned int s_ecx = reg_ecx;
-		unsigned int s_edx = reg_edx;
-		unsigned int s_esi = reg_esi;
-		unsigned int s_edi = reg_edi;
-		unsigned int s_esp = reg_esp;
-		unsigned int s_ebp = reg_ebp;
-		unsigned int s_fl  = reg_flags;
-
-		reg_eax = 0x1600/*disk change detect*/;
-		reg_dl = bios_disk;
-		CPU_SetSegGeneral(es,INT13Xfer);
-
-		imageDiskCallINT13();
-
-		if (reg_flags & FLAG_CF) {
-			if (reg_ah == 0x06) {
-				LOG(LOG_MISC,LOG_DEBUG)("INT13 image disk change flag");
-				diskChangeFlag = true;
-			}
-		}
-
-		reg_eax = s_eax;
-		reg_ebx = s_ebx;
-		reg_ecx = s_ecx;
-		reg_edx = s_edx;
-		reg_esi = s_esi;
-		reg_edi = s_edi;
-		reg_esp = s_esp;
-		reg_ebp = s_ebp;
-		reg_flags = s_fl;
-
-		busy = false;
-	}
-
-	return imageDisk::detectDiskChange();
-}
-#endif
-
-#if !defined(OSFREE)
-imageDiskINT13Drive::imageDiskINT13Drive(imageDisk *sdisk) : imageDisk(ID_INT13) {
-	subdisk = sdisk;
-	subdisk->Addref();
-
-	drvnum         = subdisk->drvnum;
-	diskname       = subdisk->diskname;
-	active         = subdisk->active;
-	sector_size    = subdisk->sector_size;
-	heads          = subdisk->heads;
-	cylinders      = subdisk->cylinders;
-	sectors        = subdisk->sectors;
-	hardDrive      = subdisk->hardDrive;
-	diskSizeK      = subdisk->diskSizeK;
-	diskChangeFlag = subdisk->diskChangeFlag;
-}
-#endif
-
-#if !defined(OSFREE)
-imageDiskINT13Drive::~imageDiskINT13Drive() {
-	subdisk->Release();
-}
-#endif
